% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cross_channel_stray_rate}
\alias{cross_channel_stray_rate}
\title{Cross Channel Stray Rate}
\format{
Vector of length 44 (4 size classes)
}
\source{
Size classes were derived from analysis of Stanislaus data
}
\usage{
cross_channel_stray_rate
}
\description{
Natural straying allocation across channels, source: expert opinion from SIT members.
}
\keyword{datasets}
