% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{month_return_proportions}
\alias{month_return_proportions}
\title{Month Return Proportions}
\format{
1 dimensional array [31 watersheds]
}
\source{
Proportion hatchery was derived using coded wire tag analysis. Details and methodology are available \href{https://cvpia-osc.github.io/fallRunDSM/articles/hatchery-analysis.html}
}
\usage{
month_return_proportions
}
\description{
The proportion of spawning fish in Oct-Dec
}
\keyword{datasets}
