% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proportion_hatchery}
\alias{proportion_hatchery}
\title{Proportion Hatchery}
\format{
1 dimensional array [31 watersheds]
}
\source{
Proportion hatchery was derived using coded wire tag analysis. Details and methodology are available \href{https://cvpia-osc.github.io/latefallRunDSM/articles/hatchery-analysis.html}{CVPIA Hatchery Analysis}
}
\usage{
proportion_hatchery
}
\description{
The proportion of hatchery fish in each tributary
}
\keyword{datasets}
