% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{route_regional_alternative}
\alias{route_regional_alternative}
\title{Route Regions - Alternative}
\source{
IP-117068
}
\usage{
route_regional_alternative(
  month,
  year,
  migrants,
  inchannel_habitat,
  floodplain_habitat,
  proportion_flow_bypass,
  detour = NULL,
  prop_pulse_flows,
  migration_survival_rate,
  territory_size,
  temperature_downstream = 19,
  density_dependent_survival,
  stochastic
)
}
\arguments{
\item{month}{The simulation month, 1-8}

\item{migrants}{An n by 4 matrix of juvenile fish by watershed and size class}

\item{inchannel_habitat}{A vector of available habitat in square meters}

\item{floodplain_habitat}{A vector of available floodplain habitat in square meters}

\item{prop_pulse_flows}{The proportion of pulse flows}

\item{migration_survival_rate}{The outmigration survival rate}

\item{territory_size}{Array of juvenile fish territory requirements for \code{\link{fill_natal}}}

\item{temperature_downstream}{Numeric temperature threshold that determines if fish move downstream. Defaults to 19.}
}
\description{
Determines if juveniles stay in the region (Sections of Mainstem
Sacramento River or San Joaquin River) or out migrate during a simulated month
}
