% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_month_dynamic.R
\name{juvenile_month_dynamic}
\alias{juvenile_month_dynamic}
\title{Juvenile Month Dynamics}
\usage{
juvenile_month_dynamic(
  hypothesis,
  fish,
  year,
  month,
  rearing_survival,
  migratory_survival,
  habitat,
  ..params,
  avg_ocean_transition_month,
  stochastic
)
}
\arguments{
\item{fish}{list object containing tracking matrices for juveniles, juveniles at chips, and adults in ocean}

\item{..params}{}
}
\description{
Juvenile Month Dynamics
}
\examples{
fish = list(juveniles = juveniles, north_delta_fish = north_delta_fish, south_delta_fish = south_delta_fish,
juveniles_at_chipps = juveniles_at_chipps, 
adults_in_ocean = adults_in_ocean)
}
