% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{late_fall_run_model}
\alias{late_fall_run_model}
\title{Late-Fall Run Chinook Model}
\source{
IP-117068
}
\usage{
late_fall_run_model(
  scenario = NULL,
  mode = c("seed", "simulate", "calibrate"),
  seeds = NULL,
  ..params = lateFallRunDSM::params
)
}
\arguments{
\item{scenario}{Model inputs, can be modified to test management actions}

\item{mode}{The mode to run model in. Can be \code{"seed"}, \code{"simulate"} or \code{"calibrate"}.}

\item{seeds}{The default value is NULL runs the model in seeding mode,
returning a 31 by 25 matrix with the first four years of seeded adults. This
returned value can be fed into the model again as the value for the seeds argument}

\item{..params}{parameters derived from calibration. More details at \code{\link{params}}}
}
\description{
Late-Fall Run Chinook life cycle model used for CVPIA's Structured
Decision Making Process
}
\examples{
fall_run_seeds <- fallRunDSM::fall_run_model(mode = "seed")
fallRunDSM::fall_run_model(scenario = DSMscenario::scenarios$ONE,
                           mode = "simulate",
                           seeds = fall_run_seeds)
}
